+++ mvd-analytics/result/messages.go
@@ -16,12 +16,12 @@ type MessagesResult struct {
 // elided via omitempty when it equals Message — consumers should treat
 // a missing MessageClean as "use Message".
 type MatchEvent struct {
-	Time         float64 `json:"time"`
-	Type         string  `json:"type"`                   // "frag", "chat", "teamsay"
-	Player       string  `json:"player"`                 // Who sent/killed
-	Team         string  `json:"team"`                   // Player's team
-	Message      string  `json:"message"`                // Chat text or frag description (markup-preserving)
-	MessageClean string  `json:"messageClean,omitempty"` // Same text with ezQuake markup stripped
-	Victim       string  `json:"victim,omitempty"`       // For frags
-	Weapon       string  `json:"weapon,omitempty"`       // For frags
+	Time         int32  `json:"time"`                   // Match-relative milliseconds (schema v8)
+	Type         string `json:"type"`                   // "frag", "chat", "teamsay"
+	Player       string `json:"player"`                 // Who sent/killed
+	Team         string `json:"team"`                   // Player's team
+	Message      string `json:"message"`                // Chat text or frag description (markup-preserving)
+	MessageClean string `json:"messageClean,omitempty"` // Same text with ezQuake markup stripped
+	Victim       string `json:"victim,omitempty"`       // For frags
+	Weapon       string `json:"weapon,omitempty"`       // For frags
 }
